+++ Projects/Plasticity/TPZLadeNelsonElasticResponse.h
@@ -425,7 +425,7 @@ inline void TPZLadeNelsonElasticResponse::
 {
   const int nVars = 6;
   int i,j;
-  typedef TFad<nVars, double> TFAD;
+  typedef TFad<nVars, REAL> TFAD;
   // extract the values of the residual vector
   
   resnorm = 0.;
